+++ internal/entities/equipment_type-entity.go
@@ -2,10 +2,9 @@ package entities
 
 import "request-system/pkg/types"
 
-
 type EquipmentType struct {
-	ID           int    `json:"id"`
-	Name         string `json:"name"`
+	ID   uint64 `json:"id"`
+	Name string `json:"name"`
 
 	types.BaseEntity
 }
